+++ frontend/src/features/authLogin/api/dto/auth.dto.ts
@@ -11,7 +11,6 @@ export interface LoginDto {
 
 export interface AuthResponse {
   access_token: string;
-  refresh_token: string;
   user: {
     id: string;
     email: string;